+++ scripts/all-in-one/utils/all_in_one_packaging.bat
@@ -5,31 +5,25 @@
 
 echo off
 IF "%1"=="" GOTO arg_not_exists
-IF "%2"=="" GOTO arg_not_exists
 set unrealVersion=%1
-set pluginFlavour=%2
 
-rmdir /s /q output
+if exist output\ rmdir /s /q output
 mkdir output
-rmdir /s /q workspace
+
+if exist workspace\ rmdir /s /q workspace
 mkdir workspace
 
 cd workspace
 
 echo ">> 1. Download wit-unreal"
-git clone https://github.com/wit-ai/wit-unreal
+git clone -c core.longpaths=true https://github.com/wit-ai/wit-unreal
 
 echo ">> 2. Download voicesdk-unreal"
-git clone https://github.com/wit-ai/voicesdk-unreal
+git clone -c core.longpaths=true https://github.com/wit-ai/voicesdk-unreal
 
 echo ">> 3. add WITH_VOICESDK macro for 'create preset' feature, this can be removed if 2 repos merged."
 node ../../utils/update_WITH_VOICESDK_flag.js
 node ../../utils/update_WITH_VOICESDK_MARKETPLACE_flag.js
-
-echo ">> 3.1 change useragent."
-node ../../update_WITH_VOICESDK_USERAGENT_flag.js
-echo ">> 3.2 change plugin flavour."
-node ../../utils/update_plugin_flavour.js %pluginFlavour%
 
 echo ">> 4. Copy code and content from wit to voicesdk"
 powershell -Command "cp ./wit-unreal/Source/* ./voicesdk-unreal/Source/ -recurse -force"
@@ -40,7 +34,17 @@ powershell -Command "Copy-Item ../../utils/Config -Destination ./voicesdk-unreal
 echo ">> 5. Build"
 if "%unrealVersion%"=="4.27" (SET vSVersion="-VS2019") else (SET vSVersion="")
 
-"C:\Program Files\Epic Games\UE_%unrealVersion%\Engine\Build\BatchFiles\RunUAT.bat" BuildPlugin -Rocket -Plugin="%CD%\voicesdk-unreal\VoiceSDK.uplugin" -TargetPlatforms=Win64 -Package="%CD%\..\output\voicesdk-unreal" %vSVersion% & cd .. & ..\\utils\\zip_all_in_one.bat
+set currDir=%CD%
+
+cd "C:\Program Files\Epic Games\UE_%unrealVersion%\Engine\Build\BatchFiles"
+
+call "RunUAT.bat" BuildPlugin -Rocket -Plugin="%currDir%\voicesdk-unreal\VoiceSDK.uplugin" -TargetPlatforms=Win64 -Package="%currDir%\..\output\voicesdk-unreal" %vSVersion%
+
+if %ERRORLEVEL% neq 0 goto:eof
+
+cd %currDir%
+cd ..
+..\\utils\\zip_all_in_one.bat
 
 goto:eof
